Replace list data structure with Map
What?
=====

After profiling token-search, it was clear a significant amount of work
was being done traversing the list of `Node`s to find matches for a
particular character.

Maps are a more efficient structure, and as a pleasant side-effect, it's
also easier to perform virtually all data management when building a
trie.

This change results in ~50% faster execution time.