Fix bug in implementation of Conduit
What?
=====

This addresses a bug introduced when switching to Conduit where tokens
were being mismatched with files.

How?

Using a ZipSource yielding filenames and chunked bytestrings resulted in
filenames becoming out of sync from the first source and the second
source, as streamed files will emit multiple chunks.

This copies over the implementation of Conduit's `sourceFile`, but
yields the pair of FilePath and ByteString, to ensure the path and
contents of the file are threaded through together. This also results in
an inability to use Conduit's built-in `decodeUtf8LenientC`, resulting
in a less-ideal implementation.

Finally, this introduces a test to ensure this works as expected.