Remove sourcing files via chunking
What?
=====

The previous implementation of `sourceFileWithFilename` yielded chunks
of files in order to reduce memory overhead (and remain inline with
behavior from Conduit).

However, the implementation of `processText` assumes the full text is
available for traversal with the trie, meaning chunked streaming of
files resulted in missed token matches. This is demonstrated via the new
long test file.

A partial solution would be to change `calculateResults` to merge
results:

    transformMap . Map.unionsWith (Map.unionWith (+)) <$>
        processAllFiles filenames newTrie

This solution would be dependent on `processTextC` handling multiple
chunks of text by carrying the list of tries between chunks of text.